# replace encoding/json with segmentio/encoding