--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,7 @@ module github.com/unrolled/render
 
 go 1.12
 
-require github.com/eknkc/amber v0.0.0-20171010120322-cdade1c07385
+require (
+	github.com/eknkc/amber v0.0.0-20171010120322-cdade1c07385
+	github.com/segmentio/encoding v0.1.3
+)