This PR replaces `encoding/json` with `segmentio/encoding`.
Here are the benchmarks from my own computer:
```
> benchcmp old.txt new.txt
benchmark                     old ns/op     new ns/op     delta
BenchmarkNormalJSON-40        1013          684           -32.48%
BenchmarkStreamingJSON-40     956           686           -28.24%
```