@unrolled 2019-12-09
Thanks for submitting the PR, but unfortunately this isn't a change I feel is needed. The goal of this package is to be a very simple rendering engine with little to no dependencies. I think if folks are worried about marshalling speeds, they wouldn't use this package and likely implement something more streamlined for their own unique needs.